Add user authentication (#7)

+++ api/files/Services/TodoService.cs
@@ -1,32 +1,31 @@
 using System.Collections.Generic;
-using Site.Config;
-using Site.Models;
-using Site.Repository;
+using Api.Config;
+using Api.Models;
 
-namespace Site.Services
+namespace Api.Services
 {
     public class TodoService : BaseService<TodoItem>
     {
         public TodoService(AppConfig config) : base(config) { }
 
-        public IEnumerable<TodoItem> GetAll() => Query($"SELECT \"Id\", \"Description\", \"Done\" FROM \"{TableName}\" ORDER BY \"Id\"");
+        public IEnumerable<TodoItem> GetAll(int userId) => Query($"SELECT \"Id\", \"Description\", \"Done\" FROM \"{TableName}\" WHERE \"UserId\"=@UserId ORDER BY \"Id\"", new { UserId = userId });
 
         public void Add(TodoItem todoItem)
         {
-            var query = $"INSERT INTO \"{TableName}\" (\"Description\", \"Done\") VALUES (@Description, @Done)";
+            var query = $"INSERT INTO \"{TableName}\" (\"Description\", \"Done\", \"UserId\") VALUES (@Description, @Done, @UserId)";
             Execute(query, todoItem);
         }
 
         public void Update(TodoItem todoItem)
         {
-            var query = $"UPDATE \"{TableName}\" SET \"Description\" = @Description, \"Done\" = @Done WHERE \"Id\" = @Id";
+            var query = $"UPDATE \"{TableName}\" SET \"Description\" = @Description, \"Done\" = @Done, \"UserId\" = @UserId WHERE \"Id\" = @Id";
             Execute(query, todoItem);
         }
 
-        public void Remove(int id)
+        public void Remove(int id, int userId)
         {
-            var query = $"DELETE FROM \"{TableName}\" WHERE \"Id\" = @Id";
-            Execute(query, new { Id = id });
+            var query = $"DELETE FROM \"{TableName}\" WHERE \"Id\" = @Id AND \"UserId\" = @UserId";
+            Execute(query, new { Id = id, UserId = userId });
         }
     }
 }
